+++ internal/registration/registration.go
@@ -87,6 +87,10 @@ func Start(ctx context.Context) error {
 		log.Fatal().Err(err).Msg("getting hostname")
 	}
 
+	if viper.GetString(keys.InstanceID) != "" {
+		hn = viper.GetString(keys.InstanceID)
+	}
+
 	if hn == "" {
 		log.Fatal().Str("hostname", hn).Msg("empty hostname")
 	}
@@ -358,4 +362,13 @@ func formatTags(tags []string) []string {
 	}
 
 	return t
+}
+
+func IsRegistered() bool {
+	if credentials.DoesFileExist(viper.GetString(keys.JwtTokenFile)) &&
+		credentials.DoesFileExist(viper.GetString(keys.ManagerIDFile)) {
+		return true
+	}
+
+	return false
 }